// BoletinCine/.stryker-tmp/sandbox8863493/src/app/shared/models/funcion/funcion.ts
// @ts-nocheck
export interface Funcion {
  idFuncion: number;
  pelicula: {
    idPelicula: string;
    nombrePelicula: string;
    imagenPortada: string;
  };
  nombreTeatro: string;
  salaFuncion: string;
  diaFuncion: string;
  horaFuncion: string;
  valorFuncion: number;
  categoriaAudioFuncion: string;
  categoriaSubtitulosFuncion: string;
  categoriaVisualFuncion: string;
}